Add functionality for persisting data across scripts
This is currently used to track the applied templates to make sure templates are not ran multiple times, but can be extended to other things as well.

Persisting data is done by storing that data in temporary files in the current working directory. These files will be removed when all templates have been applied.